+++ internal/database/online/test_impl/open.go
@@ -1,4 +1,4 @@
-package test
+package test_impl
 
 import (
 	"testing"
