+++ internal/database/online/test_impl/util.go
@@ -1,4 +1,4 @@
-package test
+package test_impl
 
 import (
 	"context"
